Allow bashrc to be used as a wrapper for a single command
Sometimes it's convenient not to have to source an environment script
into your shell if you have a lot of shells open because you can
forget which one you sourced the script in. This patch just adds a
little snippet to the end of the environment script so that if you
pass in any extra command line arguments it will treat that as a
command to run and exec it. That way you can prefix all of your
commands with the environment script instead of having to use a single
shell for building.