Build cogl-pango and all of its dependencies
The build script now additionally builds:

• fontconfig
• cairo
• pixman
• pango
• cogl-pango

Fontconfig is specially configured to look for fonts in a folder on
the Android filesystem. Apparently this works but is not documented so
it might not work forever.

Pango is configured to statically link its modules.

Maintaining the chain of PKG_CONFIG_PATH dependencies for each module
became a bit unwieldy. To simplify this after building each package it
now just always adds the prefix for that package to an exported
PKG_CONFIG_PATH variable so that it does not have to copied into each
module.